refactor: present 내 View 폴더및 파일 정리
