+++ RunningLog/Source/Present/Component/StatItem.swift
@@ -1,0 +1,25 @@
+import SwiftUI
+
+// MARK: - 통계 아이템
+struct StatItem: View {
+    let title: String
+    let value: String
+    let unit: String
+    
+    var body: some View {
+        VStack(spacing: 4) {
+            Text(title)
+                .font(.caption)
+                .foregroundColor(.secondary)
+            
+            Text(value)
+                .font(.title2)
+                .fontWeight(.semibold)
+                .foregroundColor(.primary)
+            
+            Text(unit)
+                .font(.caption2)
+                .foregroundColor(.secondary)
+        }
+    }
+}
